pyproject.toml
[project]
requires-python = ">=3.11,<4.0"
name = "finrl"
version = "0.3.8"

[tool.poetry]
description = "FinRL: 金融强化学习框架。版本0.3.5说明：稳定版本、代码重构、更多教程、清晰文档"
authors = ["Hongyang Yang, Xiaoyang Liu"]
license = "MIT"

readme = "README.md"
classifiers=[
        # Trove classifiers
        # Full list: https://pypi.python.org/pypi?%3Aaction=list_classifiers
        "License :: OSI Approved :: MIT License",
        "Programming Language :: Python :: 3.11",
        "Programming Language :: Python :: 3.12",
        "Programming Language :: Python :: Implementation :: CPython",
        "Programming Language :: Python :: Implementation :: PyPy",
    ]
keywords=["强化学习", "金融"]
[tool.poetry.urls]
github = "https://github.com/finrl/finrl-library"


[tool.poetry.dependencies]
python = "^3.11"
elegantrl = {git="https://github.com/AI4Finance-Foundation/ElegantRL.git#egg=elegantrl"}
alpaca-trade-api = "^3"
ccxt = "^3"
jqdatasdk = "^1"
pyfolio-reloaded = "^0.9"
pyportfolioopt = "^1"
ray = {extras=["default", "tune"], version = "^2"}
scikit-learn = "^1"
stable-baselines3 = {version = ">=2.0.0a5", extras = ["extra"]}
stockstats = "^0.5"
wrds = "^3"
yfinance = "^0.2"
alpaca-py = "^0.37"
webdriver-manager = "^4"
selenium = "^4"
pandas-market-calendars = "^5"


[tool.poetry.group.dev.dependencies]
black = "^24"
isort = "^5"
jupyter = "^1"
mypy = "^1"
pandas-stubs = "^2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
